compute: add CollectionEdge::flat_map_datums
`flat_map_datums(max_demand, logic)` is the canonical unified entry
point for "decoding consumers" — operators that read `Datum`s from
each row anyway (MFP, FlatMap, joins, TopK, sinks).
The closure receives a borrowed `DatumVecBorrow`, identical to the
existing `CollectionBundle::flat_map` else-branch contract.

The Vec arm reuses `DatumVec::borrow_with_limit` over each `Row` from
the timely stream.
The Columnar arm is `todo!()`; it will iterate
`Column<(Row, T, Diff)>::borrow()` via the `Rows<_>::Index` impl
(yielding `&RowRef`) and call `DatumVec::borrow_with_limit` per row,
all without materialising an owned `Row`.

Route `CollectionBundle::flat_map`'s unkeyed branch through the new
method so the per-record decode lives in exactly one place.
This deletes one of the transitional `expect_vec` fences.
