+++ src/compute/src/render/columnar.rs
@@ -34,13 +34,19 @@
 
 use columnar::Columnar;
 use differential_dataflow::VecCollection;
-use mz_repr::{Diff, Row};
+use mz_repr::{DatumVec, DatumVecBorrow, Diff, Row};
 use mz_timely_util::columnar::Column;
 use mz_timely_util::operator::CollectionExt;
-use timely::dataflow::{Scope, Stream};
+use timely::ContainerBuilder;
+use timely::dataflow::channels::pact::Pipeline;
+use timely::dataflow::operators::generic::OutputBuilder;
+use timely::dataflow::operators::generic::builder_rc::OperatorBuilder;
+use timely::dataflow::{Scope, Stream, StreamVec};
 use timely::progress::Timestamp;
 
 use crate::render::RenderTimestamp;
+use crate::render::context::{ECB, Session};
+use crate::render::errors::DataflowErrorSer;
 use crate::typedefs::KeyBatcher;
 
 /// A columnar collection of `(D, T, R)` updates traveling on a compute
@@ -172,6 +178,80 @@ impl<'scope, T: RenderTimestamp> CollectionEdge<'scope, T> {
             }
         }
         CollectionEdge::Vec(differential_dataflow::collection::concatenate(scope, vec_arm))
+    }
+
+    /// Applies `logic` to each record in this edge, exposing the record as a
+    /// borrowed [`DatumVecBorrow`] and giving it ok and err output sessions.
+    ///
+    /// `max_demand` bounds the number of columns decoded per row; pass
+    /// `usize::MAX` to decode all columns.
+    ///
+    /// This is the canonical unified entry point for "decoding consumers"
+    /// (operators that read [`mz_repr::Datum`]s from each row anyway). The
+    /// Vec arm uses [`DatumVec::borrow_with_limit`] on each [`Row`]; the
+    /// Columnar arm iterates the columnar batch directly without going
+    /// through an owned [`Row`].
+    pub fn flat_map_datums<DCB, L>(
+        self,
+        max_demand: usize,
+        mut logic: L,
+    ) -> (
+        Stream<'scope, T, DCB::Container>,
+        StreamVec<'scope, T, (DataflowErrorSer, T, Diff)>,
+    )
+    where
+        DCB: ContainerBuilder,
+        L: for<'a> FnMut(
+                &'a mut DatumVecBorrow<'_>,
+                T,
+                Diff,
+                &mut Session<T, DCB>,
+                &mut Session<T, ECB<T>>,
+            ) -> usize
+            + 'static,
+    {
+        match self {
+            CollectionEdge::Vec(c) => {
+                let scope = c.inner.scope();
+                let mut builder = OperatorBuilder::new("CollectionFlatMap".to_string(), scope);
+                let (ok_output, ok_stream) = builder.new_output();
+                let mut ok_output = OutputBuilder::<_, DCB>::from(ok_output);
+                let (err_output, err_stream) = builder.new_output();
+                let mut err_output = OutputBuilder::<_, ECB<T>>::from(err_output);
+                let mut input = builder.new_input(c.inner, Pipeline);
+                builder.build(move |_capabilities| {
+                    let mut datums = DatumVec::new();
+                    move |_frontiers| {
+                        let mut ok_output = ok_output.activate();
+                        let mut err_output = err_output.activate();
+                        input.for_each(|time, data| {
+                            // Retain the input capability to derive a `Capability` for each output;
+                            // the `Session` type alias is fixed to `Capability<T>`.
+                            let ok_cap = time.retain(0);
+                            let err_cap = time.retain(1);
+                            let mut ok_session = ok_output.session_with_builder(&ok_cap);
+                            let mut err_session = err_output.session_with_builder(&err_cap);
+                            for (v, t, d) in data.drain(..) {
+                                logic(
+                                    &mut datums.borrow_with_limit(&v, max_demand),
+                                    t,
+                                    d,
+                                    &mut ok_session,
+                                    &mut err_session,
+                                );
+                            }
+                        });
+                    }
+                });
+                (ok_stream, err_stream)
+            }
+            CollectionEdge::Columnar(_) => {
+                // Implementation will iterate `Column<(Row, T, Diff)>::borrow()`
+                // via the `Rows<_>::Index` impl (yielding `&RowRef`) and call
+                // `DatumVec::borrow_with_limit(row_ref, max_demand)` per row.
+                todo!("CollectionEdge::flat_map_datums: columnar arm")
+            }
+        }
     }
 
     /// Consolidates updates in the edge, preserving variant.
